Finalize question generator and turn it into components
