+++ leetnode/src/components/course/CourseDiscussion.tsx
@@ -35,7 +35,7 @@ import {
 
 import DateDiffCalc from "./DateDiffCalc";
 
-const Editor = dynamic(import("@/components/editor/Editor"), {
+const Editor = dynamic(import("@/components/editor/CustomRichTextEditor"), {
   ssr: false,
   loading: () => <p>Loading Editor...</p>,
 });
